refactor: streamline code structure and improve component organization across multiple files
