+++ web/src/application/root.tsx
@@ -1,16 +1,3 @@
-import { Outlet } from 'react-router';
-import { RootProvider } from '@/providers';
-import '@/index.css';
-
-export { Document as Layout } from '@/components/layouts/Document';
+export { Root as default, Document as Layout } from '@/components/Root';
 
 export const meta = () => [{ title: 'LongLink' }];
-
-/** Provides isolated Application runtime state around the active framework route. */
-export default function ApplicationRoot() {
-    return (
-        <RootProvider>
-            <Outlet />
-        </RootProvider>
-    );
-}
